requirements.txt
autopep8==1.7.0
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
pycodestyle==2.9.1
python-dotenv==0.21.0
SQLAlchemy==1.4.41
toml==0.10.2
Werkzeug==2.2.2
